Refactor tests into analyzers packages.
