+++ src/test/java/com/graphhopper/sna/analyzers/GraphAnalyzerTest.java
@@ -22,7 +22,7 @@
  * You should have received a copy of the GNU General Public License along with
  * GraphHopper-SNA. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.graphhopper.sna.centrality;
+package com.graphhopper.sna.analyzers;
 
 import com.graphhopper.sna.analyzers.GraphAnalyzer;
 import com.graphhopper.sna.data.NodeBetweennessInfo;
