+++ src/test/java/com/graphhopper/sna/analyzers/ManuallyCreatedGraphAnalyzerTest.java
@@ -22,7 +22,7 @@
  * You should have received a copy of the GNU General Public License along with
  * GraphHopper-SNA. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.graphhopper.sna.centrality;
+package com.graphhopper.sna.analyzers;
 
 import com.graphhopper.sna.model.Edge;
 import com.graphhopper.sna.model.GraphCreator;
