+++ src/test/java/com/graphhopper/sna/analyzers/examples/Graph2DAnalyzerTest.java
@@ -22,9 +22,9 @@
  * You should have received a copy of the GNU General Public License along with
  * GraphHopper-SNA. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.graphhopper.sna.centrality.examples;
+package com.graphhopper.sna.analyzers.examples;
 
-import com.graphhopper.sna.centrality.GraphAnalyzerTest;
+import com.graphhopper.sna.analyzers.GraphAnalyzerTest;
 import com.graphhopper.sna.progress.NullProgressMonitor;
 import com.graphhopper.sna.progress.ProgressMonitor;
 import java.io.FileNotFoundException;
